Big fat New Years refactor (#151)
* renamed Frameworks

(cherry picked from commit 533bbd8e6202ad903f2783683cf82fc860c26f0e)

* removed unused. fixed typo

(cherry picked from commit 79d76a358a05b0544af1859ad2eb55888e077c8a)

* extracted components from plugin-ui

* added warning icon to components

* updated tsconfig

* removing unused code

* Add code type automatically when posting code message

* Moved convert to code to its own file. also made htmlpreview a single step.

* renamed some things and removed exports from some local functions

* removed unused code in alt conversion

* altnode conversion WIP

* continuing refactor of alt node conversion. WIP

* finished refactoring alt nodes

* removed AltNode type
