+++ packages/backend/src/altNodes/altNodeUtils.ts
@@ -1,0 +1,21 @@
+import { curry } from "../common/curry";
+
+export const overrideReadonlyProperty = curry(
+  <T, K extends keyof T>(prop: K, value: any, obj: T): T =>
+    Object.defineProperty(obj, prop, {
+      value: value,
+      writable: true,
+      configurable: true,
+    }),
+);
+
+export const assignParent = overrideReadonlyProperty("parent");
+export const assignChildren = overrideReadonlyProperty("children");
+export const assignType = overrideReadonlyProperty("type");
+export const assignRectangleType = assignType("RECTANGLE");
+
+export function isNotEmpty<TValue>(
+  value: TValue | null | undefined,
+): value is TValue {
+  return value !== null && value !== undefined;
+}
