+++ packages/backend/src/index.ts
@@ -3,5 +3,5 @@ export { htmlMain } from "./html/htmlMain";
 export { tailwindMain } from "./tailwind/tailwindMain";
 export { swiftuiMain } from "./swiftui/swiftuiMain";
 export { run } from "./code";
-export { convertIntoNodes } from "./altNodes/altConversion";
+export { convertNodesToAltNodes as convertIntoNodes } from "./altNodes/altConversion";
 export * from "./messaging";
