+++ packages/backend/src/messaging.ts
@@ -11,8 +11,9 @@ export const postBackendMessage = figma.ui.postMessage;
 export const postEmptyMessage = () =>
   postBackendMessage({ type: "empty" } as EmptyMessage);
 
-export const postConversionComplete = (conversionData: ConversionMessage) =>
-  postBackendMessage(conversionData);
+export const postConversionComplete = (
+  conversionData: ConversionMessage | Omit<ConversionMessage, "type">,
+) => postBackendMessage({ ...conversionData, type: "code" });
 
 export const postError = (error: string) =>
   postBackendMessage({ type: "error", error } as ErrorMessage);
