+++ tsconfig.json
@@ -7,5 +7,5 @@
     { "path": "./apps/plugin" },
     { "path": "./apps/debug" }
   ],
-  "exclude": ["node_modules"]
+  "exclude": ["./../node_modules"]
 }
